Circuits: inline StateMachine and states
